+++ ChangeLog.md
@@ -9,8 +9,15 @@
 * Implement range/d1/d2
 * Implement max/d1/d2
 * Implement above/temperature
-* Response 405 "Method not allowed" on illigel request
+* Response 405 "Method not allowed" on illigel request (Not fully working yet)
 
 ## 0.2.0.0  -- 2018-04-26
 
 * Implement Web Routes library
+
+## 0.2.5.0  -- 2018-05-01
+
+* Adding method into WebRoutes version
+* Update cabal file
+* Fix the "response 405"
+* Improve code style